// commerceos/libs/apps/products/src/lib/products-list/enums/customer.enum.ts
export enum PeProductTypeCustomerEligibilityEnum {
  Everyone = 'EVERYONE',
  SpecificGroupsOfCustomers = 'SPECIFIC_GROUPS_OF_CUSTOMERS',
  SpecificCustomers = 'SPECIFIC_CUSTOMERS',
}
export enum PeProductArrayNamesEnum {
  Customers = 'CUSTOMERS',
  GroupsOfCustomers = 'GROUPS_OF_CUSTOMERS',
}

export enum PeCustomerFieldEnum {
  Customer = 'customerUserId',
  Group = 'customerGroupId'
}